+++ src/routes/index.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import Hero from '../components/Hero.svelte';
-  import clustersStore, { paramsStore } from '../stores/clustersStore';
+  import clustersStore from '../stores/clustersStore';
+  import { paramsStore } from '../stores/paramsStore';
 </script>
 
 {#if $clustersStore.length === 0}
@@ -10,7 +11,9 @@
     <pre data-prefix="$"><code>docker run -p 8080:8080 \</code></pre>
     <pre><code class="ml-20">-d provectuslabs/kafka-ui:latest \</code></pre>
     {#each Object.entries($paramsStore.env) as [key, value]}
-      <pre><code class="ml-20">-e {key}={value} \</code></pre>
+      {#if value != undefined}
+        <pre><code class="ml-20">-e {key}={value} \</code></pre>
+      {/if}
     {/each}
     {#if $paramsStore.javaOpts.length > 0}
       <pre><code class="ml-20">-e JAVA_OPTS=" \</code></pre>
